CURSO COMPRETO/PROJETO/requirements.txt
python-fasthtml
